+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Remove old ConfigMap and Secret when a new config version is set.
+
+### Fixed
+
 - Use `config.giantswarm.io/version` Chart annotation to determine configuration version.
 
 ## [0.2.0] - 2021-01-12